main rework for new error strategy for 1.0

+++ test/testLocalFileSmlErrorTransport.js
@@ -5,26 +5,26 @@ var expect = chai.expect;
 var mock = require('mock-require');
 mock('serialport', 'virtual-serialport');
 
-describe('test LocalFileTransport with D0Protocol', function() {
+describe('test LocalFileTransport with SMLProtocol', function() {
 
-    it('check output of two D0 messages', function(done){
+    it('check output of two SML messges', function(done){
         this.timeout(600000); // because of first install from npm
 
         var SmartmeterObis = require('../index.js');
 
         var options = {
-            'protocol': 'D0Protocol',
+            'protocol': 'SmlProtocol',
             'transport': 'LocalFileTransport',
             'transportSerialPort': '/dev/ir-usb0',
             'transportSerialBaudrate': 9600,
             'requestInterval': 10,
             'transportHttpRequestUrl': '',
             'obisNameLanguage': 'en',
-            'transportLocalFilePath': './test.d0',
-            'obisFallbackMedium': 6
+            'transportLocalFilePath': './test.sml',
+            'debug': 2
         };
 
-        var testData = '/?Bla0!\r\n6.8(0029.055*MWh)6.26(01589.28*m3)9.21(00010213)6.26*01(01563.92*m3)6.8*01(0028.086*MWh)F(0)9.20(64030874)6.35(60*m)6.6(0017.2*kW)6.6*01(0017.2*kW)6.33(001.476*m3ph)9.4(088*C&082*C)6.31(0030710*h)6.32(0000194*h)9.22(R)9.6(000&00010213&0)9.7(20000)6.32*01(0000194*h)6.36(01-01)6.33*01(001.476*m3ph)6.8.1()6.8.2()6.8.3()6.8.4()6.8.5()6.8.1*01()6.8.2*01()6.8.3*01()\r\n6.8.4*01()6.8.5*01()9.4*01(088*C&082*C)6.36.1(2013-11-28)6.36.1*01(2013-11-28)6.36.2(2016-09-24)6.36.2*01(2016-09-24)6.36.3(2015-03-26)6.36.3*01(2015-03-26)6.36.4(2013-09-27)6.36.4*01(2013-09-27)6.36.5(2000-00-00)6.36*02(01)9.36(2017-01-18&01:36:47)9.24(0.6*m3ph)9.17(0)9.18()9.19()9.25()9.1(0&1&0&-&CV&3&2.14)9.2(&&)0.0(00010213)!\r\n';
+        var testData = new Buffer('1b1b1b1b01010101760700190b4cbead6200620072630101760101070019063f3f8f0b0901454d48000041f045010163662d00760700190b4cbeae620062007263070177010b0901454d48000041f045070100620affff72620165063f2f357777078181c78203ff0101010104454d480177070100000009ff010101010b0901454d48000041f0450177070100010800ff6400018201621e52ff560009247a550177070100010801ff0101621e52ff560009247a550177070100010802ff0101621e52ff5600000000000177070100100700ff0101621b52ff55000016030177078181c78205ff0172620165063f2f3501018302e77ef33ea97bb6bba9bfa4fbd8b9f2ede51207b15acf6b98a237c21ca4982ee3ce18efe8438f1deba9d5c40eb68ae8f201010163574a00760700190b4cbeb16200620072630201710163d658000000001b1b1b1b1a03e566', 'hex');
         fs.writeFileSync(options.transportLocalFilePath, testData);
 
         var lastObisResult;
@@ -40,18 +40,19 @@ describe('test LocalFileTransport with D0Protocol', function() {
             }
             // nothing to do in this case because protocol is stateless
             expect(obisResult).to.be.an('object');
-            expect(obisResult['6-0:9.20']).to.be.an('object');
-            expect(obisResult['6-0:9.20'].rawValue).to.be.equal('64030874');
-            expect(obisResult['6-0:9.20'].values.length).to.be.equal(1);
-            expect(obisResult['6-0:9.20'].values[0].value).to.be.equal(64030874);
-            expect(obisResult['6-0:6.8']).to.be.an('object');
-            expect(obisResult['6-0:6.8'].rawValue).to.be.equal('0029.055*MWh');
-            expect(obisResult['6-0:6.8'].values.length).to.be.equal(1);
-            expect(obisResult['6-0:6.8'].values[0].value).to.be.equal(29.055);
-            expect(obisResult['6-0:6.8'].values[0].unit).to.be.equal('MWh');
+            expect(obisResult['129-129:199.130.3*255']).to.be.an('object');
+            expect(obisResult['129-129:199.130.3*255'].rawValue).to.be.empty;
+            expect(obisResult['129-129:199.130.3*255'].values.length).to.be.equal(1);
+            expect(obisResult['129-129:199.130.3*255'].values[0].value).to.be.equal('EMH');
+            expect(obisResult['1-0:1.8.1*255']).to.be.an('object');
+            expect(obisResult['1-0:1.8.1*255'].rawValue).to.be.empty;
+            expect(obisResult['1-0:1.8.1*255'].values.length).to.be.equal(1);
+            expect(obisResult['1-0:1.8.1*255'].values[0].value).to.be.equal(15338.5557);
+            expect(obisResult['1-0:1.8.1*255'].values[0].unit).to.be.equal('kWh');
 
             if (!lastObisResult) {
                 expect(counter).to.be.equal(0);
+                fs.unlinkSync(options.transportLocalFilePath);
             }
             else {
                 expect(counter).to.be.equal(1);
@@ -73,9 +74,9 @@ describe('test LocalFileTransport with D0Protocol', function() {
 
         setTimeout(function() {
             smTransport.stop();
-            expect(counter).to.be.equal(2);
-            expect(errCounter).to.be.equal(0);
-            fs.unlinkSync(options.transportLocalFilePath);
+            expect(counter).to.be.equal(1);
+            expect(errCounter).to.be.equal(1);
+            if (fs.existsSync(options.transportLocalFilePath)) fs.unlinkSync(options.transportLocalFilePath);
             done();
         }, 13000);
     });
